# api/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: PgQueryNarrative API
    description: Data Storyteller AI - Convert PostgreSQL SQL queries to business narratives
    version: v1
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /api/v1/reports/generate:
        post:
            tags:
                - reports
            summary: generate reports
            description: Generate a narrative report from a SQL query
            operationId: reports#generate
            parameters:
                - name: GenerateRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/GenerateReportPayload'
                    required:
                        - sql
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Report'
                        required:
                            - id
                            - sql
                            - narrative
                            - metrics
                            - created_at
                            - llm_model
                            - llm_provider
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/ValidationError'
                        required:
                            - name
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/LLMError'
                        required:
                            - name
                            - message
            schemes:
                - http
definitions:
    AggregateData:
        title: AggregateData
        type: object
        properties:
            avg:
                type: number
                example: 0.7402401973817502
                format: double
            count:
                type: integer
                example: 361683652
                format: int32
            max:
                type: number
                example: 0.16142315079728264
                format: double
            min:
                type: number
                example: 0.22127759169183464
                format: double
            sum:
                type: number
                example: 0.6729863868161663
                format: double
        example:
            avg: 0.03457395019421717
            count: 1823112733
            max: 0.1419711549200893
            min: 0.8860467882091754
            sum: 0.7526299824685656
    ChartSuggestion:
        title: ChartSuggestion
        type: object
        properties:
            chart_type:
                type: string
                description: 'Chart type identifier: bar, line, pie, table'
                example: Provident recusandae rerum dolor reprehenderit.
            label:
                type: string
                description: Human-readable label
                example: Alias aperiam minima eius occaecati.
            reason:
                type: string
                description: Why this chart fits the data
                example: Minima autem delectus.
        example:
            chart_type: Reprehenderit iste quidem dolor voluptates ea.
            label: Deserunt velit atque sequi.
            reason: Facere est ipsum dignissimos dicta.
        required:
            - chart_type
            - label
            - reason
...
    GenerateReportPayload:
        title: GenerateReportPayload
        type: object
        properties:
            saved_query_id:
                type: string
                example: 4ebc9d60-1da2-45e1-8a61-8e76ca341e28
                format: uuid
            sql:
                type: string
                example: "8"
                pattern: ^[^;]+$
                minLength: 1
                maxLength: 10000
        example:
            saved_query_id: 51d3a28d-adcb-4eed-b8e9-ec5c7860029b
            sql: s
        required:
            - sql
    LLMError:
        title: LLMError
        type: object
        properties:
            code:
                type: string
                example: Quos impedit ea exercitationem reiciendis.
            message:
                type: string
                example: Quos sit labore excepturi.
            name:
                type: string
                example: Perferendis esse.
        example:
            code: Vel vitae et ab.
            message: Similique veniam omnis.
            name: Nemo quidem ut qui quam.
        required:
            - name
            - message
    MetricsData:
        title: MetricsData
        type: object
        properties:
            aggregates:
                type: object
                example:
                    Omnis autem et sint est deserunt ex.:
                        avg: 0.5605439800138888
                        count: 2098515544
                        max: 0.7862947234767059
                        min: 0.9360853593719282
                        sum: 0.7603072148785976
                    Saepe nulla est voluptatibus dolores sint.:
                        avg: 0.5605439800138888
                        count: 2098515544
                        max: 0.7862947234767059
                        min: 0.9360853593719282
                        sum: 0.7603072148785976
                additionalProperties:
                    $ref: '#/definitions/AggregateData'
            period_current_label:
                type: string
                description: Label for current period when time_series is present
                example: Vero suscipit qui omnis sit veniam.
            period_previous_label:
                type: string
                description: Label for previous period
                example: Quo asperiores beatae et.
            time_series:
                type: object
                example:
                    Odit nihil.:
                        change: 0.20166130308069102
                        change_percentage: 0.7671929932026079
                        current_period: 0.6419130654630286
                        previous_period: 0.18823648238661542
                        trend: Adipisci consectetur possimus culpa ipsam.
                additionalProperties:
                    $ref: '#/definitions/TimeSeriesData'
            top_categories:
                type: object
                example:
                    Deserunt corporis error atque eveniet aliquid.:
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                    Et labore quis incidunt expedita atque voluptatibus.:
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                    Sed quaerat asperiores est sed.:
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                additionalProperties:
                    type: array
                    items:
                        $ref: '#/definitions/TopCategoryData'
                    example:
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
        example:
            aggregates:
                Commodi quo.:
                    avg: 0.5605439800138888
                    count: 2098515544
                    max: 0.7862947234767059
                    min: 0.9360853593719282
                    sum: 0.7603072148785976
                Cum sit ratione velit.:
                    avg: 0.5605439800138888
                    count: 2098515544
                    max: 0.7862947234767059
                    min: 0.9360853593719282
                    sum: 0.7603072148785976
                Nesciunt repellat vel sunt corporis.:
                    avg: 0.5605439800138888
                    count: 2098515544
                    max: 0.7862947234767059
                    min: 0.9360853593719282
                    sum: 0.7603072148785976
            period_current_label: Doloremque temporibus.
            period_previous_label: Unde molestiae odit ut.
            time_series:
                Beatae placeat quod illo ipsum id.:
                    change: 0.20166130308069102
                    change_percentage: 0.7671929932026079
                    current_period: 0.6419130654630286
                    previous_period: 0.18823648238661542
                    trend: Adipisci consectetur possimus culpa ipsam.
                Quia dolore et sint sed.:
                    change: 0.20166130308069102
                    change_percentage: 0.7671929932026079
                    current_period: 0.6419130654630286
                    previous_period: 0.18823648238661542
                    trend: Adipisci consectetur possimus culpa ipsam.
                Velit praesentium porro quia dolor aut.:
                    change: 0.20166130308069102
                    change_percentage: 0.7671929932026079
                    current_period: 0.6419130654630286
                    previous_period: 0.18823648238661542
                    trend: Adipisci consectetur possimus culpa ipsam.
            top_categories:
                Nihil non temporibus voluptatibus rerum eos.:
                    - category: Est ut quia.
                      percentage: 0.4370289696293134
                      value: 0.07511843322601428
                    - category: Est ut quia.
                      percentage: 0.4370289696293134
                      value: 0.07511843322601428
                    - category: Est ut quia.
                      percentage: 0.4370289696293134
                      value: 0.07511843322601428
    NarrativeContent:
        title: NarrativeContent
        type: object
        properties:
            drivers:
                type: array
                items:
                    type: string
                    example: Illo quis fugit ipsum corrupti sapiente.
                example:
                    - Aut illum sunt sed et.
                    - Nihil quod a in et doloremque eius.
            headline:
                type: string
                example: Odio nisi et velit quia nesciunt id.
            limitations:
                type: array
                items:
                    type: string
                    example: Quod laudantium atque dolor harum facere deleniti.
                example:
                    - Quia et debitis.
                    - Est inventore autem non.
                    - Odio repudiandae.
                    - Commodi eum.
            recommendations:
                type: array
                items:
                    type: string
                    example: Totam est rerum totam beatae et quasi.
                example:
                    - Quia perferendis atque aut eos.
                    - Beatae et quo rem natus dolor aut.
            takeaways:
                type: array
                items:
                    type: string
                    example: Libero est rerum in.
                example:
                    - Molestiae optio ut eveniet maiores.
                    - Id voluptas ea occaecati quia cum.
        example:
            drivers:
                - Fugit id voluptatem rerum quas.
                - Nobis est voluptas.
                - Qui odio laudantium.
                - Ex quos ad dolor distinctio ea.
            headline: Earum modi aut rerum molestias velit.
            limitations:
                - Non blanditiis dolore rem quam.
                - Dolorem deleniti repudiandae ab aut et.
                - Facere consequuntur eaque provident.
            recommendations:
                - Et est harum at.
                - Aliquam et eligendi sunt dolorem recusandae.
            takeaways:
                - Omnis dolores ad inventore est dolorem sed.
                - Rerum et.
                - Ex nihil harum tenetur eaque.
        required:
            - headline
            - takeaways
...
    Report:
        title: Report
        type: object
        properties:
            chart_suggestions:
                type: array
                items:
                    $ref: '#/definitions/ChartSuggestion'
                description: Suggested chart types based on result shape
                example:
                    - chart_type: Blanditiis ab id aut veniam dignissimos fugit.
                      label: Officia quod.
                      reason: Ea inventore voluptatum.
                    - chart_type: Blanditiis ab id aut veniam dignissimos fugit.
                      label: Officia quod.
                      reason: Ea inventore voluptatum.
                    - chart_type: Blanditiis ab id aut veniam dignissimos fugit.
                      label: Officia quod.
                      reason: Ea inventore voluptatum.
                    - chart_type: Blanditiis ab id aut veniam dignissimos fugit.
                      label: Officia quod.
                      reason: Ea inventore voluptatum.
            created_at:
                type: string
                example: "1986-05-19T12:46:46Z"
                format: date-time
            id:
                type: string
                example: 76e4c40c-45de-44ed-8720-4cf6c70ec90d
                format: uuid
            llm_model:
                type: string
                example: Quidem velit laboriosam magni consectetur quo beatae.
            llm_provider:
                type: string
                example: Dicta praesentium quia itaque ipsa voluptatem ad.
            metrics:
                $ref: '#/definitions/MetricsData'
            narrative:
                $ref: '#/definitions/NarrativeContent'
            saved_query_id:
                type: string
                example: fc051b9e-16a4-4017-9238-6a6b8fc0d67e
                format: uuid
            sql:
                type: string
                example: Harum molestiae perspiciatis.
        example:
            chart_suggestions:
                - chart_type: Blanditiis ab id aut veniam dignissimos fugit.
                  label: Officia quod.
                  reason: Ea inventore voluptatum.
                - chart_type: Blanditiis ab id aut veniam dignissimos fugit.
                  label: Officia quod.
                  reason: Ea inventore voluptatum.
                - chart_type: Blanditiis ab id aut veniam dignissimos fugit.
                  label: Officia quod.
                  reason: Ea inventore voluptatum.
                - chart_type: Blanditiis ab id aut veniam dignissimos fugit.
                  label: Officia quod.
                  reason: Ea inventore voluptatum.
            created_at: "2011-08-21T23:30:06Z"
            id: 85a07f91-0134-48a1-9bd3-f9c7c1cbf396
            llm_model: Quo quidem dignissimos praesentium excepturi.
            llm_provider: Illum ipsum et dolor saepe.
            metrics:
                aggregates:
                    Laboriosam dolorem.:
                        avg: 0.5605439800138888
                        count: 2098515544
                        max: 0.7862947234767059
                        min: 0.9360853593719282
                        sum: 0.7603072148785976
                period_current_label: Dolores eum non.
                period_previous_label: Et possimus quo maiores et aut.
                time_series:
                    Fugiat itaque quibusdam nisi ut.:
                        change: 0.20166130308069102
                        change_percentage: 0.7671929932026079
                        current_period: 0.6419130654630286
                        previous_period: 0.18823648238661542
                        trend: Adipisci consectetur possimus culpa ipsam.
                    Minus suscipit eum magni cupiditate illum.:
                        change: 0.20166130308069102
                        change_percentage: 0.7671929932026079
                        current_period: 0.6419130654630286
                        previous_period: 0.18823648238661542
                        trend: Adipisci consectetur possimus culpa ipsam.
                    Optio facere repellat dolorem.:
                        change: 0.20166130308069102
                        change_percentage: 0.7671929932026079
                        current_period: 0.6419130654630286
                        previous_period: 0.18823648238661542
                        trend: Adipisci consectetur possimus culpa ipsam.
                top_categories:
                    Omnis qui velit molestiae.:
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
            narrative:
                drivers:
                    - Est non et nihil officiis.
                    - Eum et vitae distinctio dolorem.
                    - Dolorum porro voluptas laudantium.
                    - Veritatis minus reiciendis laudantium.
                headline: Quae qui similique.
                limitations:
                    - Eius fugit a doloremque.
                    - Aut eveniet non sed velit.
                recommendations:
                    - Tempora et natus.
                    - Dolor quibusdam.
                    - Natus perspiciatis aperiam.
                    - Earum est explicabo.
                takeaways:
                    - Est temporibus vel maxime omnis iste.
                    - Non nesciunt incidunt illo molestias.
            saved_query_id: 32a3c3e4-2208-417d-b27b-b627d4ea4bca
            sql: Et et.
        required:
            - id
            - sql
            - narrative
            - metrics
            - created_at
            - llm_model
            - llm_provider
...
    TimeSeriesData:
        title: TimeSeriesData
        type: object
        properties:
            change:
                type: number
                example: 0.7501561209407399
                format: double
            change_percentage:
                type: number
                example: 0.6299744019172046
                format: double
            current_period:
                type: number
                example: 0.5637588210595937
                format: double
            previous_period:
                type: number
                example: 0.7508871449130278
                format: double
            trend:
                type: string
                example: Aut nam voluptatem ipsa.
        example:
            change: 0.2858033072986057
            change_percentage: 0.902448398898189
            current_period: 0.3522833087743102
            previous_period: 0.26943026802654557
            trend: Sed facere ut accusamus perspiciatis.
        required:
            - current_period
            - trend
    TopCategoryData:
        title: TopCategoryData
        type: object
        properties:
            category:
                type: string
                example: Vel enim.
            percentage:
                type: number
                example: 0.45109244937987497
                format: double
            value:
                type: number
                example: 0.6266450386190051
                format: double
        example:
            category: Sed distinctio voluptatem perferendis voluptatibus.
            percentage: 0.5723173048217615
            value: 0.8359341760367073
        required:
            - category
            - value
            - percentage
    ValidationError:
        title: ValidationError
        type: object
        properties:
            code:
                type: string
                example: Numquam aperiam cumque.
            message:
                type: string
                example: Velit consequatur officiis at porro.
            name:
                type: string
                example: Aut molestiae ut.
        example:
            code: Voluptatem doloribus.
            message: Velit dicta sit doloribus occaecati.
            name: Aliquid voluptates esse.
        required:
            - name
            - message